重构sqlite_db
